requirements.txt
djangorestframework-simplejwt
pytest
djangorestframework
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
babel==2.17.0
Brotli==1.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
convertdate==2.4.0
cryptography==45.0.7
cssselect2==0.8.0
Django==4.2.27
django-ckeditor-5==0.2.15
django-crum==0.7.9
django-js-asset==3.1.2
django-ninja==0.22.1
django-widget-tweaks==1.5.0
docxcompose==1.4.0
docxtpl==0.20.1
fonttools==4.61.1
html5lib==1.1
idna==3.10
Jinja2==3.1.6
lunardate==0.2.2
lxml==6.0.1
MarkupSafe==3.0.2
oscrypto==1.3.0
pillow==10.4.0
plum-dispatch==1.7.4
psycopg2==2.9.10
pycparser==2.23
pydantic==1.10.22
pydantic_core==2.33.2
pydyf==0.11.0
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
pyluach==2.3.0
PyMeeus==0.5.12
pypdf==6.6.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
spire-doc==13.8.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.6.3
weasyprint==66.0
webencodings==0.5.1
workalendar==17.0.0
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
pandas